ducky: move wait_for_num_versions to redpanda_installer
